Create an SSH server ilios container
This still has the PHP and Apache base, but it only runs and exposes an
SSH server to which users are added by specifying
GITHUB_ACCOUNT_SSH_USERS as a built time argument. This semi-colon
separated list takes the public keys for a users github account and
creates a login.  Each user is given passwordless sudo so they can
perform ilios administrative tasks on this machine.